Refactor: split one app related logics into separate files
