+++ src/loadenv.ts
@@ -68,8 +68,6 @@ export const naverMap = {
   apiKey: process.env.NAVER_MAP_API_KEY || "", // optional
 };
 export const oneApp = {
-  tokenSecretKey: process.env.ONEAPP_TOKEN_SECRET || "SPARCS_APP", // optional
-  accessTokenExpiry: "1h" as const, // 1시간
+  secretKey: process.env.ONEAPP_TOKEN_SECRET || "SPARCS_APP", // optional
   refreshTokenExpiry: 30 * 24 * 3600 * 1000, // 30일, ms 단위입니다.
-  ssoInfoExpiry: "5m" as const, // 5분
 };
